+++ components/pet-card/card-image.tsx
