+++ components/pet-card/card-text.tsx
