+++ types/post.ts
